Add CI/CD pipeline using GH actions
