+++ .github/workflows/build.yaml
@@ -1,0 +1,20 @@
+name: Deploy Frontend
+on: [push]
+jobs:
+  deploy-frontend:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+      - name: Setup node
+        uses: actions/setup-node@v4
+        with:
+          node-version: '23'
+      - name: Install dependencies
+        run: |
+          npm i
+          working-directory: frontend
+      - name: Build frontend
+        run: ./build.sh
+      - name: Deploy frontend
+        run: ./deploy.sh
