+++ .gitignore
@@ -4,8 +4,7 @@
 frontend/node_modules/
 frontend/.react-router/
 frontend/build/
-*.mp4
-fonts
+frontend/app/assets/
 
 # Python
 .venv
