--- frontend/app/assets/global.d.ts
@@ -1,9 +1,0 @@
-declare module '*.png' {
-  const value: any;
-  export = value;
-}
-
-declare module '*.pdf' {
-  const value: any;
-  export = value;
-}
