+++ infra/deploy/src/config/aws.py
@@ -1,0 +1,2 @@
+TERRAFORM_ENV_BUCKET_NAME = "dkavalchek-terraform"
+ASSETS_BUCKET_NAME = "dkavalchek-frontend-assets"
