--- sync_terraform.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-uv run --project ./infra/deploy ./infra/deploy/src/sync-terraform.py
