+++ sync_terraform_env.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+uv run --project ./infra/deploy ./infra/deploy/src/sync_terraform_env.py
